Add files via upload
Formatted classes.
Renamed:
CommandFunctions - ListModification
TodoAssistant1 - todo
PrintColors - ColorSelection
DirControl - UtilFunctions

ListModification has all functions that directly work on the spreadsheet
ColorSelection is the same except its not part of a class since it was redundant. Currently there are just functions in that class for colors.
todo is more or less the same
UtilFunctions has all functions that help with utility. getDirectory gets the path of Todo directory or creates it first if it doesn't exist. getWorkbook sets up a worksheet to use. cls clears the screen. checkDate checks for validity of date input.